Work around JRuby bug. Fixes #341 On JRuby, moving files using FileUtils is done via a Java call to renameTo. From the Java 6 API:  File.renameTo - Many aspects of the behavior of this method are inherently platform-dependent: The rename operation might not be able to move a file from one filesystem to another, it might not be atomic, and it might not succeed if a file with the destination abstract pathname already exists. The return value should always be checked to make sure that the rename operation was successful.
Since we only support the home directory Trash (as opposed to the  partition trash in the spec), we are allowed (as per the spec) to copy the file to the home trash and delete the original instead. We still try a direct move, first, as this is faster, especially for larger files.