Merged FindIncrementalCommand into FindNextCommand.
Added special handling of case where selection does not match query; in that
case we start search from beginning of selection, instead of after.

This is useful in the case where the "Foo" portion of "Foobar" is selected, and the
user sets the query to "Foobar". With the new handling, the selection will just expand
to all of "Foobar", rather than skipping beyond "Foo" to begin the search.