Vendor ffi-java instead of putting it in the gemspec.
This is because we don't have a version of the redcar gem for each platform,
so if the user installs redcar under 1.9.2, we still need the java ffi. And
it is not apparently possible to make a gem dependency on a gem for another
platform, which makes some sense I suppose.