Vendor json.
This is another fix to allow installation through Rubygems on any
Ruby. Starting to think that this is a misuse of Rubygems that
we should move away from soon. But anyway, Ruby 1.9.2 includes
JSON gem, but setting GEM_HOME on windows doesn't find that because
it's not a regular gem or something.