Allow plugins to be enabled/disabled.
Updated UI to show which plugins are enabled and which ones are disabled.
Also correctly handles now allowing a plugin in use by a another plugin to
be disabled unless the other one is disabled first. Also handle not allowing
plugin to be enabled if it needs another plugin that is not enabled.