Use JRuby to bundle in order to remove hard-vendored json and ffi.
By using JRuby to run the standalone bundle install, we can remove the
hard-vendored json-java and ffi-java gems from the repo. To me this
makes a lot of sense however for the time being, it relies on having
JRuby installed and running via RVM (or similar) in development in
addition to the Redcar vendored jruby-complete jar. I'm bugging people
about using MRI to install JRuby deps but it doesn't look hopeful.

If everybody hates needing two JRubies, maybe someone can rewrite the
command to use the vendored one, but that still requires getting bundler
in advance. :(