Typo error fixed.
In the Description section, there is a typo error. 
The line is written as:
"A Ruby text editor.
- written in Ruby from the ground up."

As 'A Ruby text editor' is written in the heading and all of its features are written in bullet points, so all the bullet points must follow the heading to form a complete sentence. But in the bullet point provided above, 'is' is missing.

The correct line could be written as, 
"A Ruby text editor.
- is written in Ruby from the ground up."

Thank you.